server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
motor>=3.4.0
pydantic>=2.7.4
websockets>=12.0
python-dotenv>=1.0.1
